Avoid that the contents of pub/media/catalog/product are removed when deleting an image with empty path When you have an image on a product that has an empty path (empty string, no NULL value) and you delete that image on the product edit page, the whole content of the pub/static/media folder will be removed.

+++ app/code/Magento/Catalog/Model/Product/Attribute/Backend/Media.php
@@ -341,7 +341,9 @@ class Media extends \Magento\Eav\Model\Entity\Attribute\Backend\AbstractBackend
     {
         $catalogPath = $this->_mediaConfig->getBaseMediaPath();
         foreach ($files as $filePath) {
-            $this->_mediaDirectory->delete($catalogPath . '/' . $filePath);
+            if (!empty($filePath)) {
+                $this->_mediaDirectory->delete($catalogPath . '/' . $filePath);
+            }
         }
     }
